+++ pyciemss/__init__.py
@@ -1,0 +1,1 @@
+from .mira_integration import compiled_dynamics
